issue: 4158083 Disable dpcp adapter usage for DOCA only
When DPCP flags are not defined dpcp adapter is compiled out completely.
Seperating between ibv-dpcp and DOCA devices.

Other changes:
1. Removing tap leftovers
2. Removing verify-QP. Skip non offloadable interfaces by checking DOCA interface returned names for each ib name.
3. Making ib device list set instead of a map. The mapping is not used.
