+++ src/core/dev/buffer_pool.h
@@ -82,8 +82,13 @@ public:
     void print_report(vlog_levels_t log_level = VLOG_DEBUG);
     static void print_full_report(vlog_levels_t log_level);
 
+#ifdef DEFINED_DPCP_PATH_RX_OR_TX
     uint32_t find_lkey_by_ib_ctx_thread_safe(ib_ctx_handler *p_ib_ctx_h);
+#endif // DEFINED_DPCP_PATH_RX_OR_TX
+#ifndef DEFINED_DPCP_PATH_RX_AND_TX
     doca_mmap *get_doca_mmap() const { return m_allocator_data.get_doca_mmap(); }
+#endif // !DEFINED_DPCP_PATH_RX_AND_TX
+
     /**
      * Get buffers from the pool - thread safe
      * @parma pDeque List to put the buffers.
