+++ src/core/dev/doca/hw_queue_tx_doca.h
@@ -78,7 +78,7 @@ public:
     void down();
     int modify_qp_ratelimit(struct xlio_rate_limit_t &, uint32_t) { return -1; }
     void dm_release_data(mem_buf_desc_t *) {}
-    uint32_t is_ratelimit_change(struct xlio_rate_limit_t &) { return 0;}
+    uint32_t is_ratelimit_change(struct xlio_rate_limit_t &) { return 0; }
     void ti_released(xlio_ti *) override {} // Dummy
     doca_notification_handle_t get_notification_handle() const { return m_notification_handle; }
     uint32_t send_doca_single(void *ptr, uint32_t len, mem_buf_desc_t *user_data);
