+++ src/core/sock/sockinfo_tcp.cpp
@@ -4722,6 +4722,7 @@ int sockinfo_tcp::getsockopt_offload(int __level, int __optname, void *__optval,
                 if (m_p_connected_dst_entry) {
                     ring *tx_ring = m_p_connected_dst_entry->get_ring();
                     if (tx_ring) {
+#ifdef DEFINED_DPCP_PATH_RX_OR_TX
                         /*
                          * For bonding we get context of the 1st slave. This approach
                          * works for RoCE LAG mode.
@@ -4730,9 +4731,13 @@ int sockinfo_tcp::getsockopt_offload(int __level, int __optname, void *__optval,
                         if (p_ib_ctx_h) {
                             struct xlio_pd_attr *pd_attr = (struct xlio_pd_attr *)__optval;
                             pd_attr->flags = 0;
-                            pd_attr->ib_pd = (void *)p_ib_ctx_h->get_ibv_pd();
+                            pd_attr->ib_pd = (void *)p_ib_ctx_h->get_ctx_ibv_dev().get_ibv_pd();
                             ret = 0;
                         }
+#else // DEFINED_DPCP_PATH_RX_OR_TX
+                        ret = -1;
+                        errno = EPERM;
+#endif // DEFINED_DPCP_PATH_RX_OR_TX
                     }
                 }
             }
